+++ opentelemetry-sdk/CHANGELOG.md
@@ -6,6 +6,8 @@
   ([#751](https://github.com/open-telemetry/opentelemetry-python/pull/751))
 - Rename Measure to ValueRecorder in metrics
   ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))
+- Add SumObserver in metrics
+  ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))
 
 ## 0.8b0